Changed iex to only require one file included
