+++ iex/test/company_test.cc
@@ -4,8 +4,6 @@
  * @copyright 2020 Antony Kellermann
  */
 
-#include "iex/api/company.h"
-
 #include <gtest/gtest.h>
 
 #include "iex/iex.h"
