+++ iex/test/symbols_test.cc
@@ -4,8 +4,6 @@
  * @copyright 2020 Antony Kellermann
  */
 
-#include "iex/api/symbols.h"
-
 #include <gtest/gtest.h>
 
 #include <vector>
