+++ iex/test/system_status_test.cc
@@ -4,8 +4,6 @@
  * @copyright 2020 Antony Kellermann
  */
 
-#include "iex/api/system_status.h"
-
 #include <gtest/gtest.h>
 
 #include "iex/iex.h"
